+++ py/server/deephaven/server/__init__.py
@@ -1,6 +1,3 @@
 #
 # Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
 #
-
-# Packages under the deephaven.server heading are not meant to be called externally - it exists as a convenient place
-# for the server to execute implementation logic via python
